+++ core/components/ajaxupload/docs/changelog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.5] - 2023-03-10
+
+### Added
+
+- Add scriptTpl and uploadSectionTpl property for AjaxUpload snippet
+- Add ajaxupload.image_tpl system setting.
+
+### Fixed
+
+- Fix ajaxuploadAllowOverwrite property in AjaxUpload2FormIt hook
+- Fix Default lexicon entry is not set in AjaxUploadRequired [#62]
+
 ## [1.6.4] - 2023-01-24
 
 ### Fixed